Refactor StudyRoomsPreview to call API from container with the rest of them
